requirements.txt
ffmpeg-python==0.2.0
matplotlib==3.5.2
numpy==1.23.1
Pillow==9.2.0
PySide2==5.15.2.1
pyvista==0.35.2
pyvistaqt==0.9.0
PyYAML==6.0